(modules) controller added for vehicles and readings
